Updated state time and weight to unsigned long long, to accommodate large routes.

+++ core/fibheap/fibheap.h
@@ -42,7 +42,7 @@ Boston, MA 02110-1301, USA.  */
 
 #include "stdio.h"
 
-typedef long fibheapkey_t;
+typedef unsigned long long fibheapkey_t;
 //typedef double fibheapkey_t;
 
 typedef struct fibheap
